Feat: tags shown on all posts, seed with post tags
Added a post tags part of the seeding process to make testing easier.
Gets the created posts from before and just calls the `addTags` method
on them for the set of tags we want that post to have.

Also create a bootstrap `Badge` per tag per post on the all posts page.