Display and filter by tags on gallery page
Get all tags on will mount and when that is done add them to the
component state.  Display these tags that are in the state.  When one is
clicked on then set that tag in the state as the filter tag.  The shown
posts are first filted down based on the filter tag.